Release v15.2.0 - Multiple dungeon starts & performance improvements

+++ app/Service/DungeonRoute/Logging/ThumbnailServiceLoggingInterface.php
@@ -19,7 +19,7 @@ interface ThumbnailServiceLoggingInterface
         ?int   $viewportHeight,
         ?int   $imageWidth,
         ?int   $imageHeight,
-        ?int   $zoomLevel,
+        ?float $zoomLevel,
         ?int   $quality,
     ): void;
 
@@ -33,7 +33,7 @@ interface ThumbnailServiceLoggingInterface
         ?int   $viewportHeight,
         ?int   $imageWidth,
         ?int   $imageHeight,
-        ?int   $zoomLevel,
+        ?float $zoomLevel,
         ?int   $quality,
     ): void;
